Prevents to created files on download error
It prevents to create a file when an error rise while downloading a
book. Other wise the user will be prompted to type a HTTP error 404
page or other alike.

When this happen 'speed-type' exits with a message error on the
minibuffer telling the user about a download error with a given book number.