66 Improve Testing, Refactoring and Documentation
Add Docmentation
+ Fix most flyway-errors to improve doc-string
+ Add doc-string to most buffer-local vars
+ Add uniline flow-diagram and describe each phase
+ Adjust "running speed-type" to include newly added features

Consolidate `speed-type--setup-code` and `speed-type--code-with-highlighting`
using `speed-type--setup`

Make argument-order of `speed-type--setup` consistent

Always return `speed-type-buffer` when calling `speed-type--setup`

Return nil when reading stats and no statistic-file is present. This
is necessary for last-continue-at to be 0 and no error is thrown.

Rename `speed-type-gb-dir` to `speed-type-directory` and introduce
alias. The directory is used for all kinds of retrieved content and stats not
only gutenberg (gb).

Replace speed-type--fill-region with it's content.

Increment version-number to 1.5

Add various tests:
+ reading and saving stats
+ `speed-type-region` with prog-mode, fundamental-mode and hexl-mode (file-ref)
+ downcase
+ point-movement
+ time-register: pause/resume
+ stop-words-p
+ find-last-continue-at-point-in-stats